ptbcontrib/ptb_jobstores/requirements.txt
python-telegram-bot[job-queue]~=20
pymongo>=4.1,<5
SQLAlchemy==1.4.46
